Rewrite structure
